Enable verification of token database
When verification fails, the user either needs to start
with -reindex-tokens manually, or let the wallet reindex itself.
Around the upgrade to 5.0, a full reindex is a sane default.

- Rename -reindextokens to -reindex-tokens
- Move ReindexTokenDB after setting chainActive
- Add VerifyTokenDB
- Add block hash of token creation to db
- Do not buffer tokens when reindexing

+++ src/tokens/tokengroupmanager.cpp
@@ -89,7 +89,7 @@ void CTokenGroupManager::ResetTokenGroups() {
     CTransaction tgTxWgr;
     CTokenGroupDescription tgDescriptionWGR("WGR", "Wgr", 8, "https://wagerr.com", uint256());
     CTokenGroupStatus tokenGroupStatus;
-    CTokenGroupCreation tgCreationWGR(MakeTransactionRef(tgTxWgr), tgInfoWGR, tgDescriptionWGR, tokenGroupStatus);
+    CTokenGroupCreation tgCreationWGR(MakeTransactionRef(tgTxWgr), uint256(), tgInfoWGR, tgDescriptionWGR, tokenGroupStatus);
     mapTokenGroups.insert(std::pair<CTokenGroupID, CTokenGroupCreation>(NoGroup, tgCreationWGR));
 
 }
